+++ Games/The Grand Reserve/js/main.js
@@ -1,6 +1,6 @@
 // --- MODULE ENTRY POINT ---
 import {initFirebase, openSyncModal, closeSyncModal, copySyncCode, loadRemoteSyncCode, triggerManualSync} from "./firebase-sync.js";
-import {startLoop, plantSeed, addToPress, removeFromPress, placeWineOnRack, removeWineFromRack, addToKettle, removeFromKettle, startKettlePhysics, adjustKettleHeat, clearFinishedKettle, sellWineQualityGroup, sellBeerGroup, buySeed, buyPantryItem, buyPlot, buyBarrel, buyKettle, buyOakConditioning, saveCustomLabel, bottleAsVinegar, setWeather, startWeatherSystem} from "./engine.js";
+import {startLoop, plantSeed, addToPress, removeFromPress, placeWineOnRack, removeWineFromRack, addToKettle, removeFromKettle, startKettlePhysics, adjustKettleHeat, clearFinishedKettle, sellWineQualityGroup, sellBeerGroup, buySeed, buyPantryItem, buyPlot, buyBarrelType, buyKettle, buyOakConditioning, saveCustomLabel, bottleAsVinegar, setWeather, startWeatherSystem} from "./engine.js";
 import {initSound, switchReserveTab, switchShopTab, switchTab, updateHeaderUI, renderPlots, closePlotSelector, renderCellarUI, openPressModal, closePressModal, renderWarehouse, renderMarket, openSellModal, openSellBeerModal, closeSellModal, renderShop, renderBreweryUI, renderRacks, openRackSelectModal, closeRackSelectModal, openKettleModal, closeKettleModal, closeModal, openLabelerModal, closeLabelerModal, updateLabelDraft, renderWeather} from "./ui.js";
 import { initPlaytest } from "./playtest.js";
 import { globals } from "./state.js";
@@ -19,7 +19,7 @@ window.removeFromKettle = removeFromKettle;
 window.closeModal = closeModal;
 window.closeSellModal = closeSellModal;
 window.buyPlot = buyPlot;
-window.buyBarrel = buyBarrel;
+window.buyBarrelType = buyBarrelType;
 window.buyKettle = buyKettle;
 window.buyOakConditioning = buyOakConditioning;
 window.buySeed = buySeed;